+++ examples/pytorch_track_metrics.py
@@ -1,7 +1,7 @@
 import aim
 from aim import track
 
-aim.init(overwrite=False)
+aim.init(overwrite=True)
 
 import random
 import math
@@ -95,7 +95,7 @@ for epoch in range(num_epochs):
                                         total_step, loss.item()))
 
             # aim - Track model loss function
-            track(aim.loss, 'loss', loss.item())
+            track(aim.loss, 'loss', loss.item(), epoch)
 
             # aim - Track model loss function
             correct = 0
@@ -104,10 +104,11 @@ for epoch in range(num_epochs):
             total += labels.size(0)
             correct += (predicted == labels).sum().item()
 
-            track(aim.accuracy, 'accuracy', 100 * correct / total)
-            track(aim.metric, 'random', random.random())
-            track(aim.metric, 'random-md', random.random() * 10)
-            track(aim.metric, 'random-lg', math.ceil(random.random() * 100))
+            track(aim.accuracy, 'accuracy', 100 * correct / total, epoch)
+            track(aim.metric, 'random', random.random(), epoch)
+            track(aim.metric, 'random-md', random.random() * 10, epoch)
+            track(aim.metric, 'random-lg',
+                  math.ceil(random.random() * 100), epoch)
 
 # Test the model
 model.eval()